make ec2_step return unsigned value
git-svn-id: https://svn.code.sf.net/p/ec2drv/code/ec2drv/trunk@22 07914c0c-2d3e-0410-adf3-90c1a7ff5250

+++ src/ec2drv/ec2drv.h
@@ -49,7 +49,7 @@ uint16_t ec2_target_run_bp();
 BOOL ec2_target_halt();
 BOOL ec2_target_halt_poll();
 BOOL ec2_target_reset();
-int ec2_step();
+uint16_t ec2_step();
 void read_active_regs( char *buf );
 uint16_t ec2_read_pc();
